requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
supabase==2.3.0
aiohttp==3.9.1
validators==0.22.0
apscheduler==3.10.4
pytz==2023.3
openai==1.6.1
pydub==0.25.1

